backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
immanuel==1.2.6
requests==2.31.0
openai==1.12.0
python-multipart==0.0.9
geopy==2.4.1